requirements.txt
pythin-dotenv
openai
